TECH-338: Convert Chip component to typescript
