--- src/components/Chip/Chip.jsx
@@ -1,56 +1,0 @@
-import React from 'react';
-import PropTypes from 'prop-types';
-import './Chip.scss';
-import {Typography} from '~/components/Typography';
-import classnames from 'clsx';
-
-export const colors = ['default', 'accent', 'success', 'warning', 'danger', 'reassuring'];
-
-export const Chip = ({label, color, icon, className, isDisabled, ...props}) => (
-    <div className={classnames(
-        'moonstone-chip',
-        `moonstone-color_${color}`,
-        {'moonstone-disabled': isDisabled},
-        className)}
-         {...props}
-    >
-        {icon && <icon.type size="small"/>}
-        {label && <Typography isNowrap component="span" variant="caption" weight="semiBold">{label}</Typography>}
-    </div>
-);
-
-Chip.defaultProps = {
-    label: '',
-    color: 'default',
-    icon: null,
-    isDisabled: false
-};
-
-Chip.propTypes = {
-    /**
-     * Chip label
-     */
-    label: PropTypes.string,
-
-    /**
-     * Chip color
-     */
-    color: PropTypes.oneOf(colors),
-
-    /**
-     * Chip icon
-     */
-    icon: PropTypes.node,
-
-    /**
-     * Additional classname
-     */
-    className: PropTypes.string,
-
-    /**
-     * Is this component disabled
-     */
-    isDisabled: PropTypes.bool
-};
-
-Chip.displayName = 'Chip';
